Unionize packets and improve mutex syntax

+++ rust/src/sdk/mcu/random.rs
@@ -3,5 +3,5 @@ use crate::sdk::mcu::register::read_reg_rnd_number;
 
 #[inline(always)]
 pub fn rand() -> u16 {
-    return (clock_time() as u16 ^ read_reg_rnd_number()) as u16;
+    return clock_time() as u16 ^ read_reg_rnd_number();
 }
